waypoint command reorganization, small fixes
